pyproject.toml
[project]
name = "self-claw"
version = "0.6.0"
description = "轻量级企业 Agent 框架"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.30,<1",
    "pydantic>=2.7,<3",
    "httpx>=0.27,<1",
    "apscheduler>=3.10,<4",
    "openai>=1.30,<2",
    "anthropic>=0.30,<1",
    "chromadb>=0.5,<1",
    "python-dotenv>=1.0,<2",
    "structlog>=24.1,<26",
    "tiktoken>=0.7,<1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.23,<1",
    "pytest-cov>=5,<6",
    "ruff>=0.5,<1",
    "mypy>=1.10,<2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
